% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{bvMoranI}
\alias{bvMoranI}
\title{Compute aggregation score of each cellular module based on Global Bivariate Moran's I}
\usage{
bvMoranI(coords, st_fq, ref)
}
\arguments{
\item{coords}{The coordinates of each spot}

\item{st_fq}{The frequencies of each cell subset in each spot.}

\item{ref}{The "ref$filter" object enables us to obtain the components of each cellular module.}
}
\value{
A data frame containing the aggregation score of each cellular module.
}
\description{
Compute aggregation score of each cellular module based on Global Bivariate Moran's I
}
\examples{
bvMoranI(coords,st_fq,ref)
}
