% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{cm_network}
\alias{cm_network}
\title{Construct cellular network}
\usage{
cm_network(NMFres, Corres, ...)
}
\arguments{
\item{NMFres}{Co-occurring nodes in each CM obtained from NMF results.}

\item{Corres}{Specifically correlated edges obtained from correlation results.}

\item{top_n}{Select the upper limit of the number of cell subsets in each CM.}

\item{corr}{Filter specifically correlated edges with correlation greater than corr.}

\item{fdr}{Filter specifically correlated edges with pval_fdr less than fdr.}
}
\value{
A list contains the following parts."global": the global network, "each": the network of each CM, "raw": NMF result, "filter": the components of each CM, "ann": the correspondence between subClusters and majorClusters. It can be used to draw networks with the functions "gr.igraph_each" and "gr.igraph_global", export the components of CMs, or serve as a reference to recover cellular modules in Tutorial 2.
}
\description{
Interconnect co-occurring nodes via specifically correlated edges to construct CM networks.
}
\examples{
cm_network(NMF_K12,cor_pair)
cm_network(NMF_K12,cor_pair,top_n=10,corr=0.2,fdr=0.05)
}
