% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{colocalization}
\alias{colocalization}
\title{Compute colocalization score of each cellular module based on Spearman correlation coefficients}
\usage{
colocalization(st_fq, ref, ...)
}
\arguments{
\item{st_fq}{The frequencies of each cell subset in each spot.}

\item{ref}{The "ref$filter" object enables us to obtain the components of each cellular module.}

\item{method}{Methods for calculating correlation:"pearson" and "spearman". Default is "spearman".}
}
\value{
A data frame containing the colocalization score of each cellular module.
}
\description{
Compute colocalization score of each cellular module based on Spearman correlation coefficients
}
\examples{
colocalization(st_fq,ref,method="spearman")
}
