% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{creNet}
\alias{creNet}
\title{Create and Visualize Multicellular Networks}
\usage{
creNet(node, edge, n = 10)
}
\arguments{
\item{node}{data frame of nodes of each network}

\item{edge}{data frame of edges of networks}

\item{n}{number of nodes}
}
\value{
plot networks
}
\description{
Create and visualize multicellular networks
}
