% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{freq_calculate}
\alias{freq_calculate}
\title{Calculate raw cell subset frequency matrix}
\usage{
freq_calculate(meta)
}
\arguments{
\item{meta}{Cell annotation dataframe (The dataframe should contain at least three columns: "sampleID", "majorCluster","subCluster").}
}
\value{
Raw cell subsets frequency matrix.
}
\description{
Calculate frequency matrix from cell annotation dataframe.
}
\examples{
freq_calculate(meta)
}
