% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{freq_normalize}
\alias{freq_normalize}
\title{Normalize cell subset frequency matrix}
\usage{
freq_normalize(mat_fq, ...)
}
\arguments{
\item{mat_fq}{Raw cell subset frequency matrix.}

\item{normalize}{Normalize methods:"minmax" represents Min-Max Normalization,"zscore" represents Z-Score Normalization.}
}
\value{
Normalized cell subset frequency matrix.
}
\description{
Normalize cell subset frequency matrix
}
\examples{
freq_normalize(mat_fq)
freq_normalize(mat_fq,normalize="minmax")
}
