% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{freq_normalize_st}
\alias{freq_normalize_st}
\title{Normalize cell subset frequency matrix in spatial data}
\usage{
freq_normalize_st(spe, ref, ...)
}
\arguments{
\item{spe}{A seurat object(spatial transcriptomics data). The meta data should contain annotation information for each subCluster and be consistent with the reference, representing the abundance of each cell subset in each spot.}

\item{ref}{The "ref$ann" object contains all the subCluster and corresponding majorCluster information, so that we can compute the frequencies of cell subsets within corresponding broad cell types.}

\item{normalize}{Normalize methods:"minmax" represents Min-Max Normalization,"zscore" represents Z-Score Normalization.}
}
\value{
A seurat object(normalized spatial transcriptomics data).
}
\description{
Normalize cell subset frequency matrix in spatial data
}
\examples{
freq_normalize_st(spe,ref,normalize="minmax")
}
