% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{freq_reduction}
\alias{freq_reduction}
\title{Perform PHATE using the frequencies of cell subsets within a specific CM}
\usage{
freq_reduction(meta, subsets)
}
\arguments{
\item{meta}{A data frame containing cell frequencies and other supplementary information.}

\item{subsets}{The cell subsets within a specific CM.}
}
\value{
An AnnData object after PHATE reduction.
}
\description{
Perform PHATE using the frequencies of cell subsets within a specific CM
}
\examples{
freq_reduction(meta,subsets)
}
