% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getEdge}
\alias{getEdge}
\title{Get Edges of networks}
\usage{
getEdge(mat)
}
\arguments{
\item{mat}{frequency matrix}
}
\value{
data frame of all potential edges of networks.
}
\description{
Get all specifically correlated cell-subset pairs as potential edges of networks
}
