% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNode}
\alias{getNode}
\title{Get Nodes of Networks}
\usage{
getNode(mat, K)
}
\arguments{
\item{mat}{frequency matrix}

\item{K}{number of CMs}
}
\value{
data frame of all potential nodes of each network.
}
\description{
get all potential nodes of each network.
}
