% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.igraph_each}
\alias{gr.igraph_each}
\title{Visualize the network of each CM}
\usage{
gr.igraph_each(each, ...)
}
\arguments{
\item{each}{The result of function "cm_network".}

\item{Layout}{The layout methods in igraph (Default is layout_in_circle).}
}
\value{
A graph of cellular network of each CM.
}
\description{
Visualize the network of each CM
}
\examples{
gr.igraph_each(each)
gr.igraph_each(each,Layout=layout_in_circle)
}
