% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.igraph_global}
\alias{gr.igraph_global}
\title{Visualize the global network}
\usage{
gr.igraph_global(global, ...)
}
\arguments{
\item{global}{The result of function "cm_network".}

\item{Layout}{The layout methods in igraph (Default is layout_in_circle).}
}
\value{
A graph of the global network.
}
\description{
Visualize the global network
}
\examples{
gr.igraph_global(global)
gr.igraph_global(global,Layout=layout_with_fr)
}
