% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.proximity}
\alias{gr.proximity}
\title{Calculate the spatial proximity between cell subset pairs within several specified cellular modules.}
\usage{
gr.proximity(coords, st_fq, ref, cm_select)
}
\arguments{
\item{coords}{The coordinates of each spot.}

\item{st_fq}{The frequencies of each cell subset in each spot.}

\item{ref}{The "ref$filter" object enables us to obtain the components of each cellular module.}

\item{cm_select}{A character containing the names of specified cellular modules. For example: c("CM01","CM02").}
}
\value{
A heatmap displaying the spatial proximity between cell subset pairs.
}
\description{
Calculate the spatial proximity between cell subset pairs within several specified cellular modules.
}
\examples{
gr.proximity(coords,st_fq,c("CM01","CM02"))
}
