% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.trajectory}
\alias{gr.trajectory}
\title{A heatmap displaying changes along the sample trajectory, including cell frequencies, CM activity and other supplementary information}
\usage{
gr.trajectory(data, subset, rowann)
}
\arguments{
\item{data}{An Anndata object after PHATE and Palantir analysis.}

\item{subset}{The cell subsets to be displayed.}

\item{rowann}{Select other features as annotations for the heatmap.}
}
\value{
A heatmap displaying the sample trajectory.
}
\description{
A heatmap displaying changes along the sample trajectory, including cell frequencies, CM activity and other supplementary information
}
\examples{
gr.trajectory(data, subset, rowann)
}
