% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.weight_all}
\alias{gr.weight_all}
\title{Visualize the weights of all cell subsets in CMs}
\usage{
gr.weight_all(nmf_res)
}
\arguments{
\item{nmf_res}{The result of function nmf, which includes the feature matrix.}
}
\value{
A heatmap displaying the weights of all cell subsets within each cellular module.
}
\description{
Visualize the weights of all cell subsets in CMs
}
\examples{
gr.weight_all(nmf_res)
}
