% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gr.weight_top}
\alias{gr.weight_top}
\title{Visualize cell subsets with the highest weights in each CM}
\usage{
gr.weight_top(nmf_res, ...)
}
\arguments{
\item{nmf_res}{The result of function nmf, which includes the feature matrix.}

\item{num}{The number of cell subsets with the highest weights in each CM.}
}
\value{
The top cell subsets sorted by NMF weights for each cellular module.
}
\description{
Visualize cell subsets with the highest weights in each CM
}
\examples{
gr.weight_top(nmf_res,num=15)
}
