% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{pair_correlation}
\alias{pair_correlation}
\title{Calculate correlation coefficients between cell subsets}
\usage{
pair_correlation(mat_fq, ...)
}
\arguments{
\item{mat_fq}{Raw cell subset frequency matrix.}

\item{method}{Methods for calculating correlation:"pearson" or "spearman". Default is "pearson".}
}
\value{
A data frame containing correlation between all pairs of cell subsets, including the following columns: "subCluster1", "subCluster2", "majorCluster1", "majorCluster2", "correlation", "pval", "pval_fdr", "spe". "spe" is the specificity of subset pairs. For each element 𝑟,where i<j in cor_pair, its background (complementary) set is defined as {rik,rkj|k!=i,j}. The specificity is defined as the fraction of elements in the background set that not exceeding rij.
}
\description{
Calculate correlation coefficients between cell subsets
}
\examples{
pair_correlation(mat_fq)
pair_correlation(mat_fq,method="pearson")
}
