% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{pseudotime}
\alias{pseudotime}
\title{Perform trajectory inference using Palantir}
\usage{
pseudotime(data, root)
}
\arguments{
\item{data}{An AnnData object.}

\item{root}{Select a cluster as the starting point.}
}
\value{
An AnnData object after trajectory inference.
}
\description{
Perform trajectory inference using Palantir
}
\examples{
pseudotime(data,root)
}
