% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{sc_cm_recover}
\alias{sc_cm_recover}
\title{Recover cellular modulrs in scRNA_seq data}
\usage{
sc_cm_recover(ref, mat_fq)
}
\arguments{
\item{ref}{The parameter "ref" is the result of the function cm_network, which serves as the reference file of cellular modules recovery.}

\item{mat_fq}{Normalized sample-cell frequency matrix.}
}
\value{
A matrix containing the abundance of the CMs in each sample.
}
\description{
Utilize the reference file to recover the sample CMT in scRNA_seq data.
}
\examples{
sc_cm_recover(ref,mat_fq)
}
