% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVarNet.R
\name{st_cm_recover}
\alias{st_cm_recover}
\title{Recover cellular modules in spatial transcriptomics data}
\usage{
st_cm_recover(weight, spe)
}
\arguments{
\item{weight}{The weights of the cell subsets that make up each cellular module.}

\item{spe}{A seurat object(normalized spatial transcriptomics data) containing the frequencies of each cell subset in each spot.}
}
\value{
A seurat object containg the abundance of each cellular module in each spot.
}
\description{
Recover cellular modules in spatial transcriptomics data
}
\examples{
st_cm_recover(weight,spe)
}
